move some logic back into main to make init process clearer and reduce scope of configs
