+++ config/feature.js
@@ -1,8 +1,5 @@
 "use strict";
 
-const DevToolsUtils = require("devtools-sham/shared/DevToolsUtils");
-const AppConstants = require("devtools-sham/sham/appconstants").AppConstants;
-
 let config = {};
 
 const developmentConfig = require("./development.json");
@@ -46,50 +43,15 @@ function isDevelopment() {
   return isEnabled("development");
 }
 
-function isDevToolsPanel() {
-  return process.env.NODE_ENV === "DEVTOOLS_PANEL";
-}
-
 function getPath(path, parentObj) {
   const keys = path.split(".");
   return keys.reduce((obj, key) => obj[key], parentObj);
 }
 
-function getTargetFromQuery() {
-  const href = window.location.href;
-  const nodeMatch = href.match(/ws=([^&#]*)/);
-  const firefoxMatch = href.match(/firefox-tab=([^&#]*)/);
-  const chromeMatch = href.match(/chrome-tab=([^&#]*)/);
-
-  if (nodeMatch) {
-    return { type: "node", param: nodeMatch[1] };
-  } else if (firefoxMatch) {
-    return { type: "firefox", param: firefoxMatch[1] };
-  } else if (chromeMatch) {
-    return { type: "chrome", param: chromeMatch[1] };
-  }
-
-  return null;
-}
-
-function setConfigs() {
-  // Set various flags before requiring app code.
-  if (isEnabled("clientLogging")) {
-    DevToolsUtils.dumpn.wantLogging = true;
-  }
-
-  if (isEnabled("development")) {
-    AppConstants.DEBUG_JS_MODULES = true;
-  }
-}
-
 module.exports = {
   isEnabled,
   getValue,
-  getTargetFromQuery,
-  setConfigs,
   isDevelopment,
-  isDevToolsPanel,
   setConfig,
   getConfig
 };
