+++ public/js/main.js
@@ -6,11 +6,18 @@ const { Provider } = require("react-redux");
 const ReactDOM = require("react-dom");
 const React = require("react");
 
-const {
-  getTargetFromQuery, setConfigs, isDevToolsPanel
-} = require("../../config/feature");
+const DevToolsUtils = require("devtools-sham/shared/DevToolsUtils");
+const AppConstants = require("devtools-sham/sham/appconstants").AppConstants;
+const { isEnabled } = require("../../config/feature");
 
-setConfigs();
+// Set various flags before requiring app code.
+if (isEnabled("clientLogging")) {
+  DevToolsUtils.dumpn.wantLogging = true;
+}
+
+if (isEnabled("development")) {
+  AppConstants.DEBUG_JS_MODULES = true;
+}
 
 const configureStore = require("./util/create-store");
 const reducers = require("./reducers");
@@ -52,12 +59,29 @@ function renderRoot(component) {
   );
 }
 
+function getTargetFromQuery() {
+  const href = window.location.href;
+  const nodeMatch = href.match(/ws=([^&#]*)/);
+  const firefoxMatch = href.match(/firefox-tab=([^&#]*)/);
+  const chromeMatch = href.match(/chrome-tab=([^&#]*)/);
+
+  if (nodeMatch) {
+    return { type: "node", param: nodeMatch[1] };
+  } else if (firefoxMatch) {
+    return { type: "firefox", param: firefoxMatch[1] };
+  } else if (chromeMatch) {
+    return { type: "chrome", param: chromeMatch[1] };
+  }
+
+  return null;
+}
+
 const connTarget = getTargetFromQuery();
 if (connTarget) {
   startDebugging(connTarget, boundActions).then(() => {
     renderRoot(App);
   });
-} else if (isDevToolsPanel()) {
+} else if (process.env.NODE_ENV === "DEVTOOLS_PANEL") {
   // The toolbox already provides the tab to debug.
   module.exports = {
     setThreadClient: firefox.setThreadClient,
